Add general helper class Util
Right now this just contains Util::isAdmin, which verifies (based on the
application configuration) whether a given hostmask is an administrator
or not. Currently, this is not integrated with the plugins due to the
fact that I am not sure how to extract the entire hostmask for a sender
when handling an event.